requirements.txt
future==0.18.2
python-whois==0.7.3
XlsxWriter==1.3.7
